refactor: move collectors config to config/python/ subdirectory
Prepares for future C collector config under config/c/; each backend
subdirectory owns its loader, config/utils.py reduced to instantiate().
